6.增加了紧急播出单查看的级别对应信息的功能

+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/SysUserMapper.java
@@ -120,7 +120,7 @@ public interface SysUserMapper
      * @param userId 用户ID
      * @return 用户对象信息
      */
-    public String selectAid(Long userId);
+    public String selectAid(int userId);
 
     /**
      * 根据用户ID查询用户等级roleid
@@ -128,5 +128,5 @@ public interface SysUserMapper
      * @param userId 用户ID
      * @return 用户对象信息
      */
-    public int selectRoleid(Long userId);
+    public int selectRoleid(int userId);
 }
